pyproject.toml
[project]
name = "claude-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.66.0",
    "python-dotenv>=1.1.1",
]
